+++ src/matrices_functions.cpp
@@ -23,9 +23,20 @@ std::vector<double> seq_num2(double start, double end, double step){
     values.push_back(cumul);
   }
   return values;
-
 }
 
+// a simple function to create a vector of values between a start and an end with defined step
+// [[Rcpp::export]]
+std::vector<int> seq_num2f(int start, int end, int step){
+
+  std::vector<int> values;
+  int cumul = 0 - step;
+  while(cumul+step <= end){
+    cumul+=step;
+    values.push_back(cumul);
+  }
+  return values;
+}
 
 
 // a simple function to find the index of the first occurence of value in a numeric vector