+++ src/matrices_functions.h
@@ -5,6 +5,8 @@ std::vector<double> seq_num(double start, double end, double step);
 
 std::vector<double> seq_num2(double start, double end, double step);
 
+std::vector<int> seq_num2f(int start, int end, int step);
+
 int get_first_index(NumericVector& v1, double x);
 
 std::vector<int> get_all_indeces(NumericVector& v1, double x);